requirements.txt
# Dynamic Pricing Model Requirements - Production Grade MLOps System

# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.1.0
joblib>=1.2.0

# Web API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Enhanced Features
holidays>=0.34  # For holiday detection
prophet>=1.1.4  # For demand forecasting

# MLOps and Experiment Tracking
mlflow>=2.8.0

# Data visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0

# API integration
requests>=2.28.0

# Development and testing
pytest>=7.0.0
jupyter>=1.0.0
